key value storage with bytes
