+++ sercli/src/server/db_storage.rs
@@ -4,7 +4,7 @@ use sqlx::{Executor, PgPool, query};
 pub struct DBStorage {}
 
 impl DBStorage {
-    pub async fn set(key: &str, value: &str, pool: &PgPool) -> Result<()> {
+    pub async fn set(key: &str, data: &[u8], pool: &PgPool) -> Result<()> {
         Self::create_table(pool).await?;
 
         pool.execute(
@@ -16,7 +16,7 @@ impl DBStorage {
 ",
             )
             .bind(key)
-            .bind(value),
+            .bind(data),
         )
         .await
         .map_err(|e| anyhow!(e))?;
@@ -24,22 +24,31 @@ impl DBStorage {
         Ok(())
     }
 
-    pub async fn get(key: &str, pool: &PgPool) -> Result<Option<String>> {
+    pub async fn get(key: &str, pool: &PgPool) -> Result<Option<Vec<u8>>> {
         Self::create_table(pool).await?;
 
-        let result: Option<(String,)> = sqlx::query_as("SELECT value FROM key_value_storage WHERE key = $1;")
-            .bind(key)
-            .fetch_optional(pool)
-            .await?;
+        let result: Option<(Vec<u8>,)> =
+            sqlx::query_as("SELECT value FROM key_value_storage WHERE key = $1;")
+                .bind(key)
+                .fetch_optional(pool)
+                .await?;
 
         Ok(result.map(|(value,)| value))
+    }
+
+    pub async fn set_str(key: &str, val: &str, pool: &PgPool) -> Result<()> {
+        Self::set(key, val.as_bytes(), pool).await
+    }
+
+    pub async fn get_str(key: &str, pool: &PgPool) -> Result<Option<String>> {
+        Ok(Self::get(key, pool).await?.map(|vec| String::from_utf8(vec).unwrap()))
     }
 
     async fn create_table(pool: &PgPool) -> Result<()> {
         pool.execute(query(
             r"CREATE TABLE IF NOT EXISTS key_value_storage (
               key VARCHAR(255) PRIMARY KEY,
-              value VARCHAR(255)
+              value BYTEA NOT NULL
 );",
         ))
         .await
@@ -53,6 +62,7 @@ impl DBStorage {
 mod test {
 
     use anyhow::Result;
+    use fake::{Fake, Faker};
     use sqlx::Executor;
 
     use crate::{db::prepare_db, server::db_storage::DBStorage};
@@ -65,13 +75,27 @@ mod test {
 
         assert_eq!(DBStorage::get("sokol", &pool).await?, None);
 
-        DBStorage::set("sokol", "sobaka", &pool).await?;
+        DBStorage::set_str("sokol", "sobaka", &pool).await?;
 
-        assert_eq!(DBStorage::get("sokol", &pool).await?, Some("sobaka".to_string()));
+        assert_eq!(
+            DBStorage::get_str("sokol", &pool).await?,
+            Some("sobaka".to_string())
+        );
 
-        DBStorage::set("sokol", "boran", &pool).await?;
+        DBStorage::set_str("sokol", "boran", &pool).await?;
 
-        assert_eq!(DBStorage::get("sokol", &pool).await?, Some("boran".to_string()));
+        assert_eq!(
+            DBStorage::get_str("sokol", &pool).await?,
+            Some("boran".to_string())
+        );
+
+        for _ in 0..100 {
+            let buff: Vec<u8> = Faker.fake();
+
+            DBStorage::set("buff", &buff, &pool).await?;
+
+            assert_eq!(DBStorage::get("buff", &pool).await?, Some(buff));
+        }
 
         Ok(())
     }
